requirements.txt
astroid==2.11.5
async-timeout==4.0.2
autopep8==1.6.0
botocore==1.24.21
certifi==2022.5.18.1
charset-normalizer==2.0.12
Deprecated==1.2.13
dill==0.3.5.1
idna==3.3
isort==5.10.1
jmespath==1.0.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
packaging==21.3
platformdirs==2.5.2
protobuf==3.19.4
pyasn1==0.4.8
pycodestyle==2.8.0
pylint==2.14.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
redis==4.3.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.37
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
